fixing unittests for types in structarr

+++ vsm/structarr.py
@@ -83,9 +83,9 @@ def enum_array(arr, indices=[], field_name='i'):
           dtype=[('i', '<i8'), ('value', '<i8')])
     """
     if len(indices) == 0:
-        indices = np.arange(arr.size)
+        indices = np.arange(arr.size, dtype=np.int)
     else:
-        indices = np.array(indices)
+        indices = np.array(indices, dtype=np.int)
     return zip_arr(indices, arr, field_names=[field_name, 'value'])
 
 
@@ -124,10 +124,10 @@ def enum_matrix(arr, axis=0, indices=[], field_name='i'):
     #if len(indices) == 0:
     #    indices = np.arange(arr.shape[1])
     if len(indices) == 0 and len(arr.shape) > 1:
-        indices = np.arange(arr.shape[1])
+        indices = np.arange(arr.shape[1], dtype=np.int)
     
     if type(indices) == list:
-        indices = np.array(indices)
+        indices = np.array(indices, dtype=np.int)
     ind = np.array([indices.copy() for i in xrange(arr.shape[0])])
     dt = [(field_name, indices.dtype), ('value', arr.dtype)]
     mt = zip_arr(ind, arr, field_names=[field_name, 'value'])
@@ -185,9 +185,9 @@ def enum_sort(arr, indices=[], field_name='i', filter_nan=False):
     """
     idx = np.argsort(arr)
     if len(indices) == 0:
-        indices = np.arange(arr.shape[0])
+        indices = np.arange(arr.shape[0], dtype=np.int)
     else:
-        indices = np.array(indices)
+        indices = np.array(indices, dtype=np.int)
 	
     dt = [(field_name, indices.dtype), ('value', arr.dtype)]
